major refactor of sync scripts
