+++ docs/sync/1.1-lecture.md
@@ -1,0 +1,8 @@
+---
+title: Lecture
+custom_edit_url: https://github.com/cryptoeconomics-study/code//blob/master/ch1/1.1/lecture.md
+hide_title: false
+---
+<!-- This file is generated by /website/scripts/sync-util.js - changes will be overwritten! -->
+
+404: Not Found
