--- website/scripts/sync-getting-started.js
@@ -1,19 +1,0 @@
-const { syncPages } = require('./sync-util')
-
-const GIT_REF = 'master'
-const REPO = 'cryptoeconomics-study-dao/code/'
-
-// note: for the sync scripts to work the contentLocation fields are CASE SENSITIVE
-
-const pages = [
-	{
-		destination: 'docs/sync/development-setup.md',
-		title: 'Development Setup',
-		contentLocation: 'README.md'
-	}
-]
-
-const locationReferenceMap = {}
-
-syncPages(pages, locationReferenceMap, GIT_REF, REPO)
-
