+++ website/sidebars.json
@@ -1,9 +1,9 @@
 {
   "docs": {
     "Getting Started": [
-      "getting-started/welcome",
+      "sync/welcome",
       "sync/development-setup",
-			"getting-started/course-overview"
+			"sync/course-overview"
     ],
 		"Ch1: Payment Processor": [
 			"ch1/index",
@@ -11,7 +11,7 @@
         "type": "subcategory",
         "label": "1.1 Hashes and Signatures",
         "ids": [
-          "ch1/1.1/lecture",
+          "sync/1.1-lecture",
           "sync/1.1-code-challenge"
         ]
       },
@@ -19,7 +19,7 @@
         "type": "subcategory",
         "label": "1.2 Payment Processor",
         "ids": [
-          "ch1/1.2/lecture",
+          "sync/1.2-lecture",
           "sync/1.2-code-challenge"
         ]
       },
@@ -27,7 +27,7 @@
         "type": "subcategory",
         "label": "1.3 Replay Protection",
         "ids": [
-          "ch1/1.3/lecture",
+          "sync/1.3-lecture",
           "sync/1.3-code-challenge"
         ]
       },
@@ -35,7 +35,7 @@
         "type": "subcategory",
         "label": "1.4 Account Model vs UTXOs",
         "ids": [
-          "ch1/1.4/lecture",
+          "sync/1.4-lecture",
           "sync/1.4-code-challenge"
         ]
       },
@@ -43,7 +43,7 @@
         "type": "subcategory",
         "label": "1.5 Centralized Systems",
         "ids": [
-          "ch1/1.5/lecture",
+          "sync/1.5-lecture",
           "sync/1.5-code-challenge"
         ]
       }
@@ -133,12 +133,12 @@
   },
   "dao": {
     "Cryptoeconomics.Study DAO": [
-      "dao/welcome",
-      "dao/gerbils",
-			"dao/carrots",
-			"dao/advisers",
-			"dao/help-wanted",
-			"dao/code-of-conduct"
+      "sync/dao-welcome",
+      "sync/dao-gerbils",
+			"sync/dao-carrots",
+			"sync/dao-advisers",
+			"sync/dao-help-wanted",
+			"sync/dao-code-of-conduct"
     ]
 	}
 }
